+++ src/ForgotPassword/index.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import './index.css';
-import { useRequestPasswordReset } from "../hooks/useRequestPassword";
+import { useRequestPasswordReset } from "../hooks/useResetPassword";
