+++ src/utils/ fetchReset.js
@@ -1,5 +1,4 @@
-const API_BASE_URL =
-  process.env.REACT_APP_API_BASE_URL || "https://greensmtaani-d6ee50db917a.herokuapp.com";
+const API_BASE_URL = process.env.REACT_APP_API_BASE_URL || "https://greensmtaani-d6ee50db917a.herokuapp.com";
 
 async function fetchData(endpoint, method = "GET", body = null) {
   const options = {
@@ -9,16 +8,19 @@ async function fetchData(endpoint, method = "GET", body = null) {
   if (body) options.body = JSON.stringify(body);
 
   const response = await fetch(`${API_BASE_URL}${endpoint}`, options);
+
   if (!response.ok) {
-    let errorMessage = `Error: ${response.status}`;
-    try {
-      const errorData = await response.json();
-      errorMessage = errorData.detail || errorMessage;
-    } catch {}
+    const errorData = await response.json();
+    const errorMessage = errorData.detail || `Error: ${response.status}`;
     throw new Error(errorMessage);
   }
+
   return response.json();
 }
 
 export const requestPasswordReset = (email) =>
   fetchData("/reset-request/", "POST", { email });
+
+export const resetPassword = (email, otp, password) =>
+  fetchData("/reset-password/", "PUT", { email, otp, password });
+
